backend/requirements.txt
fastapi
uvicorn[standard]
python-jose[cryptography]
passlib[bcrypt]
SQLAlchemy
psycopg2-binary
pydantic
pydantic-settings
python-dotenv

